refactor: colocate recorder reference panel
